refactor:[BE]Review DDD 리팩토링
